+++ addin/AutoLoadPyxelRestAddIn/Updater.cs
@@ -9,29 +9,28 @@ namespace AutoLoadPyxelRestAddIn
     {
         private static readonly ILog Log = LogManager.GetLogger("Updater");
         
-        private readonly string pythonPath;
-        private readonly string pathToUpdateScript;
+        private readonly string updateScriptPath;
         private readonly string pathToUpToDateConfigurations;
         private readonly bool checkPreReleases;
 
         internal Updater()
         {
-            pythonPath = ThisAddIn.GetSetting("PathToPython");
+            string pythonPath = ThisAddIn.GetSetting("PathToPython");
             if (!File.Exists(pythonPath))
                 throw new Exception(string.Format("Path to Python '{0}' cannot be found.", pythonPath));
 
-            pathToUpdateScript = ThisAddIn.GetSetting("PathToUpdateScript");
-            if (!File.Exists(pathToUpdateScript))
-                throw new Exception(string.Format("PyxelRest auto update script '{0}' cannot be found.", pathToUpdateScript));
+            updateScriptPath = Path.Combine(Path.GetDirectoryName(pythonPath), "pyxelrest_auto_update.exe");
+            if (!File.Exists(updateScriptPath))
+                throw new Exception(string.Format("PyxelRest auto update script '{0}' cannot be found.", updateScriptPath));
 
             pathToUpToDateConfigurations = ThisAddIn.GetSetting("PathToUpToDateConfigurations");
-            if (!Boolean.TryParse(ThisAddIn.GetSetting("CheckPreReleases"), out checkPreReleases))
+            if (!bool.TryParse(ThisAddIn.GetSetting("CheckPreReleases"), out checkPreReleases))
                 checkPreReleases = false; // Do not check for pre-releases by default
         }
 
         internal void CheckUpdate()
         {
-            string commandLine = this.pathToUpdateScript;
+            string commandLine = string.Empty;
             if (!string.IsNullOrEmpty(pathToUpToDateConfigurations))
                 commandLine += string.Format(" --path_to_up_to_date_configurations {0}", pathToUpToDateConfigurations);
             if (checkPreReleases)
@@ -39,7 +38,7 @@ namespace AutoLoadPyxelRestAddIn
 
             Log.Debug("Check for PyxelRest update.");
             Process updateScript = new Process();
-            updateScript.StartInfo.FileName = pythonPath;
+            updateScript.StartInfo.FileName = updateScriptPath;
             updateScript.StartInfo.Arguments = commandLine;
             updateScript.StartInfo.UseShellExecute = false;
             updateScript.StartInfo.CreateNoWindow = true;